requirements.txt
gym
-e git+https://gitlab.com/awarelab/gym-sokoban.git#egg=gym-sokoban
Pillow
graphviz